refactor: rename GameSessionActiveWithNarrative to ActiveGameSession
- Simplified component name from GameSessionActiveWithNarrative to ActiveGameSession
- Updated all imports and references
- Renamed test file to match
- Added debug logging to track narrative segment generation
